Refactor of the main components from the app.
