+++ react-app/src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 
 import './scss/index.scss';
-import Game from '../src/app/components/Game'; // eslint-disable-line import/first
+import Game from '../src/app/screens/Game'; // eslint-disable-line import/first
 
 import registerServiceWorker from './registerServiceWorker';
